Move compute_intersections into own file
