+++ src/irregular/shape_element_intersections.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "packingsolver/irregular/shape.hpp"
+
+namespace packingsolver
+{
+namespace irregular
+{
+
+std::vector<Point> compute_intersections(
+        const ShapeElement& element_1,
+        const ShapeElement& element_2);
+
+}
+}
