+++ test/irregular/CMakeLists.txt
@@ -2,6 +2,7 @@ add_executable(PackingSolver_irregular_test)
 target_sources(PackingSolver_irregular_test PRIVATE
     trapezoid_test.cpp
     shape_test.cpp
+    shape_element_intersections_test.cpp
     shape_convex_hull_test.cpp
     shape_extract_borders_test.cpp
     shape_self_intersections_removal_test.cpp
